packages/runtime/pyproject.toml
[project]
name = "swael-runtime"
version = "1.0.2"
description = "SwarmKit runtime: topology interpreter, LangGraph compiler, AGT-backed governance, CLI and HTTP server."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "Srijith Kartha" }]
keywords = ["agents", "multi-agent", "swarm", "langgraph", "mcp", "agt"]
classifiers = [
    "Development Status :: 2 - Pre-Alpha",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

# Dependencies below are intentionally un-pinned placeholders. Lock concrete
# versions once Phase 1 implementation begins — see design §20.1.
dependencies = [
    # Core
    "pydantic>=2.8",
    "pyyaml>=6.0",
    "typer>=0.12",           # CLI — `swarmkit` entrypoint
    "rich>=13.7",            # terminal rendering for authoring chat
    "httpx>=0.27",
    "anyio>=4.4",

    # Execution engine (design §7, §14)
    "langgraph>=0.2",
    "langchain-core>=0.3",

    # Governance — Microsoft AGT (design §8, §16).
    # Pinned to 3.x; AGT is iterating fast (public preview, v1→v3 in 7 weeks).
    # The GovernanceProvider ABC (§8.5) isolates the runtime from AGT API churn.
    "agent-os-kernel>=3.2.0,<4.0",
    "agentmesh-platform>=3.2.0,<4.0",
    "agent-sre>=3.2.0,<4.0",

    # MCP (design §18)
    "mcp>=1.0",

    # LLM SDKs — one per ModelProvider built-in (design/details/model-provider-abstraction.md).
    # Only model_providers/ imports these; everything else goes through ModelProvider ABC.
    "anthropic>=0.39",
    "google-genai>=1.0",
    "openai>=1.50",
    # Ollama uses httpx only (already in core deps) — no dedicated SDK.

    # HTTP server for persistent mode (design §14.1)
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",

    # Storage / checkpointing (design §14.5)
    "sqlalchemy>=2.0",
    # "psycopg[binary]>=3.2",  # optional Postgres backend

    # Schema package (workspace sibling)
    "swael-schema",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=5.0",
    "mypy>=1.11",
    "ruff>=0.6",
    "types-pyyaml",
]

[project.scripts]
swael = "swael_runtime.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/swael_runtime"]
